+++ src/index.css
@@ -5,27 +5,36 @@
 :root {
   --fa-display: block;
 }
-html,
-body,
-#root {
-  height: 100%;
-}
 
-.btn {
-  @apply text-white font-medium py-2.5 px-5 rounded-lg text-sm min-w-[105px] text-center focus:ring-4 focus:outline-none active:scale-95 transition-transform duration-200;
-}
-.btn-primary {
-  @apply bg-gradient-to-tr from-fuchsia-500 via-violet-600 to-blue-400 hover:bg-gradient-to-tl focus:ring-violet-300;
-}
+@layer base {
+  html,
+  body,
+  #root {
+    height: 100%;
+    -webkit-tap-highlight-color: transparent;
+  }
 
-.input {
-  @apply h-11 w-full rounded-xl border border-transparent bg-gray-200 pl-3 pr-3 focus:bg-transparent focus:ring-violet-500;
+  body {
+    @apply text-gray-700;
+  }
 }
+@layer components {
+  .btn {
+    @apply text-white font-medium py-2.5 px-5 rounded-lg text-sm min-w-button text-center focus:ring-4 focus:outline-none active:scale-95 transition-transform duration-200;
+  }
+  .btn-primary {
+    @apply bg-gradient-to-tr from-fuchsia-500 via-violet-600 to-blue-400 hover:bg-gradient-to-tl focus:ring-violet-300;
+  }
 
-.input-error {
-  @apply border-red-500;
-}
+  .input {
+    @apply h-11 w-full rounded-xl border border-transparent bg-gray-200 pl-3 pr-3 focus:bg-transparent focus:ring-violet-500;
+  }
 
-.text-error {
-  @apply mt-1 text-xs font-medium tracking-wide text-red-500;
+  .input-error {
+    @apply border-red-500;
+  }
+
+  .text-error {
+    @apply mt-1 text-xs font-medium tracking-wide text-red-500;
+  }
 }
